refactor: move budgeting::project_budget from lib to wire
